fix(ci): corregir warnings de flutter analyze que bloqueaban CI
- main.dart: eliminar import unused 'flutter/services.dart'
- project_backup_service.dart: eliminar imports unused (dart:convert, dart:typed_data, http, path_provider alias); cast innecesario
- backup_viewmodel.dart: eliminar comparación null innecesaria (tipo no nullable)
- catalog_viewmodel.dart: eliminar variable stackTrace no usada en catch
- file_viewer_view.dart: eliminar campo unused _resolvedLinuxMediaUrl; eliminar ! innecesario en bestCandidate
- admin_backups_view.dart: marcar _pickSavePath con ignore (lógica real, temporalmente desconectada)
- flutter-ci.yml: crear .env placeholder antes de analyze/build (fix asset_does_not_exist)

+++ lib/services/project_backup_service.dart
@@ -1,11 +1,7 @@
 import 'dart:io';
-import 'dart:convert';
-import 'dart:typed_data';
 import 'package:archive/archive.dart';
 import 'package:path/path.dart' as path;
-import 'package:http/http.dart' as http;
 import 'package:path_provider/path_provider.dart';
-import 'package:path_provider/path_provider.dart' as path_provider;
 
 /// Información de un backup del proyecto
 class ProjectBackupInfo {
@@ -614,7 +610,7 @@ class ProjectBackupService {
                 fileBytes,
               );
               archive.addFile(archiveFile);
-              totalSize = (totalSize as int) + fileBytes.length;
+              totalSize = totalSize + fileBytes.length;
               filesAdded++;
               print('   ✅ Añadido: $relativePath (${fileBytes.length} bytes)');
             } catch (e) {
